design_generation_service/requirements.txt
# Third-party libraries
fastapi
uvicorn
gunicorn
httpx
pydantic
google-cloud-storage
python-pptx